Renamed src to lib
